% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_debug.R
\name{create_stat_with_caching}
\alias{create_stat_with_caching}
\title{Create a debug version of the given Stat}
\usage{
create_stat_with_caching(OriginalStat, methods = c("parameters",
  "aesthetics", "setup_params", "setup_data", "compute_layer",
  "compute_panel", "compute_group", "finish_layer"), verbose = FALSE,
  head_n = 6)
}
\arguments{
\item{OriginalStat}{Ggproto Stat object to debug. e.g. StatIdentity}

\item{methods}{Methods within ggproto object to debug. Default: all methods}

\item{verbose}{Print arguments and return values during ggplot build and render process. Default: FALSE}

\item{head_n}{Argument to \code{head()} when printing data.frames. Default: 6}
}
\value{
A subclass of the given Stat with each method modified to capture
the argument list and return value.
}
\description{
Create a debug version of the given Stat by creating a new Stat which directly
inherits from the original Stat.  Then for each named method, rewrite the
method body to capture arguments and return values for each call.
}
