% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_data_cache}
\alias{get_data_cache}
\title{Get the cached data}
\usage{
get_data_cache(nested = TRUE)
}
\arguments{
\item{nested}{Nest elements from the same call under a single name. Default: TRUE.
Set to FALSE for a flatter list structure.}
}
\value{
list of values from the last run
}
\description{
Get the cached data
}
