% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbench.R
\name{ggbench}
\alias{ggbench}
\title{Benchmark ggplot2 plots}
\usage{
ggbench(exprs = ggbench_standards(1), ..., device = withr::with_png,
  filename = tempfile(fileext = "Rplot\%03d"))
}
\arguments{
\item{exprs}{A \code{\link[rlang:quos]{rlang::quos()}} of ggplot specs like that generated by
\code{\link[=ggbench_standards]{ggbench_standards()}}.}

\item{...}{Passed to \link[bench:mark]{bench::mark}}

\item{device}{A device wraper that takes \code{filename} (see \link[withr:with_png]{withr::with_png})
as its first argument.}

\item{filename}{A filename, possibly using \code{\link[=sprintf]{sprintf()}} syntax to render more than
one plot.}
}
\value{
A \code{\link[bench:mark]{bench::mark()}} object.
}
\description{
Benchmark ggplot2 plots
}
\examples{
ggbench()

}
