% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbench.R
\name{ggbench_standards}
\alias{ggbench_standards}
\title{Standard list of ggplot benchmarks}
\usage{
ggbench_standards(which = 1)
}
\arguments{
\item{which}{A subset of the standards, or \code{NULL} for everything}
}
\value{
A named \code{\link[rlang:quos]{rlang::quos()}} of plot expressions
}
\description{
Standard list of ggplot benchmarks
}
\examples{
ggbench_standards()

}
