% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{ggcheckout}
\alias{ggcheckout}
\title{Checkout a repo and replace a directory with its contents}
\usage{
ggcheckout(repo = "tidyverse/ggplot2", path = tempfile(),
  ignore = c("\\\\.Renviron", "\\\\.Rprofile"))
}
\arguments{
\item{repo}{The repo from which to fetch ggplot2 source}

\item{path}{The path that should be replaced}

\item{ignore}{Files to ignore when removing all files in the target directory}
}
\description{
This is intended for use with RStudio cloud to test different versions of ggplot2
with different versions of R.
}
