% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdebug-package.R
\docType{package}
\name{ggdebug-package}
\alias{ggdebug}
\alias{ggdebug-package}
\title{ggdebug: Debug functions and ggproto methods in ggplot2}
\description{
Debugs functions and ggproto methods in ggplot2.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paleolimbot/ggdebug}
  \item Report bugs at \url{https://github.com/paleolimbot/ggdebug/issues}
}

}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (0000-0002-9415-4582)

}
\keyword{internal}
