% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{ggdebug}
\alias{ggdebug}
\alias{ggundebug}
\alias{ggundebug_all}
\title{Debug ggplot2 functions}
\usage{
ggdebug(...)

ggundebug(...)

ggundebug_all(...)
}
\arguments{
\item{...}{One or more functions or \code{\link[ggplot2:ggproto]{ggplot2::ggproto()}} methods (see \code{\link[=ggdebug_one]{ggdebug_one()}})}
}
\description{
Debugs functions or \code{\link[ggplot2:ggproto]{ggplot2::ggproto()}} methods in ggplot2. Expressions are
evaluated in the ggplot2 namespace, so they can be unexported functions as well.
Use \code{undebug_all()} to clear all debugged functions. Note that ggplot2 need not
be attached to use these functions.
}
\examples{
library(ggplot2)
# the problem is in ggplot() or GeomPoint$draw_panel()
ggdebug(ggplot, GeomPoint$draw_panel)
# ggplot(mpg, aes(cty, hwy)) + geom_point()
# definitely not in ggplot(), but need to take another look
ggundebug(ggplot)
# ggplot(mpg, aes(cty, hwy)) + geom_point()
# stop debugging
ggundebug_all()

}
