% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gginfo}
\alias{gginfo}
\alias{ggrepo}
\alias{ggurl}
\alias{ggbrowse}
\alias{ggrefer}
\alias{ggpath}
\alias{gginstall}
\title{Information about the currently installed version of ggplot2}
\usage{
gginfo(include_base = FALSE, dependencies = FALSE)

ggrepo()

ggurl(path = "/", repo = ggrepo())

ggbrowse(path = "/", repo = ggrepo())

ggrefer(path = "/", repo = ggrepo())

ggpath(path = "/")

gginstall(repo = "tidyverse/ggplot2", ...,
  INSTALL_opts = "--with-keep.source")
}
\arguments{
\item{include_base, dependencies}{See \code{\link[sessioninfo:package_info]{sessioninfo::package_info()}}}

\item{path}{A path to a file within ggplot2}

\item{repo}{A repository and reference, like \code{tidyverse/ggplot2} or \code{tidyverse/ggplot2@v3.1.1}.}

\item{INSTALL_opts, ...}{Options for install. Using --with-keep.source allows
the functions in this section to find the GitHub repo from whence
ggplot2 came.}
}
\description{
Information about the currently installed version of ggplot2
}
\examples{
# install a specific version from GitHub with source attached
# gginstall("tidyverse/ggplot2")

# get information about the currently installed version
gginfo()
ggrepo()

# browse the source on GitHub
ggbrowse()
ggbrowse("R/aes.r")

# can also pass functions like in ggdebug()
ggbrowse(aes)
ggbrowse(FacetWrap$map_data)

}
