% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggmethods}
\alias{ggmethods}
\title{Get all methods of a ggproto object}
\usage{
ggmethods(obj, ns = getNamespace("ggplot2"), ignore = "super",
  include_super = TRUE)
}
\arguments{
\item{obj}{A ggplot2 ggproto object.}

\item{ignore}{Method names to ignore}

\item{include_super}{Follow superclass ggproto objects}
}
\value{
A quosures that can be spliced (\code{!!!}) into \code{\link[=ggdebug]{ggdebug()}} or \code{\link[=ggtrace]{ggtrace()}}
}
\description{
Get all methods of a ggproto object
}
