% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{ggtrace}
\alias{ggtrace}
\title{Trace ggplot2 functions}
\usage{
ggtrace(..., on_enter = quo(NULL), on_exit = quo(NULL))
}
\arguments{
\item{...}{One or more functions or \code{\link[ggplot2:ggproto]{ggplot2::ggproto()}} methods (see \code{\link[=ggtrace_one]{ggtrace_one()}})}

\item{on_enter}{A \code{\link[rlang:quo]{rlang::quo()}} to evaluate on entering the \code{fun}.}

\item{on_exit}{A \code{\link[rlang:quo]{rlang::quo()}} to evaluate on exiting the \code{fun}.}
}
\description{
Traces functions or \code{\link[ggplot2:ggproto]{ggplot2::ggproto()}} methods in ggplot2. Expressions are
evaluated in the ggplot2 namespace, so they can be unexported functions as well.
Use \code{gguntrace_all()} to clear all traced functions. Note that ggplot2 need not
be attached to use these functions. These functions do not use
\code{\link[=trace]{trace()}} and \code{\link[=untrace]{untrace()}}, because these functions do not work with ggproto methods.
}
