% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_inject.R
\name{inject_data_into_stat_function}
\alias{inject_data_into_stat_function}
\title{Inject an argument or override the return value of a method within a Stat}
\usage{
inject_data_into_stat_function(OriginalStat, method, args = NULL,
  return_value = NULL)
}
\arguments{
\item{OriginalStat}{the actual Stat}

\item{method}{the method name to inject data into}

\item{args}{named list of values to insert at the beginning of the method.
Usually to override arguments. Default: NULL (i.e. no injection)}

\item{return_value}{the value to return from the method (overriding the actual value).
Default: NULL (i.e. no injection)}
}
\description{
Inject an argument or override the return value of a method within a Stat
}
