% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nest_list}
\alias{nest_list}
\title{Nest list elements of the same name under a single list entry}
\usage{
nest_list(ll)
}
\arguments{
\item{ll}{list}
}
\value{
nested list
}
\description{
Nest list elements of the same name under a single list entry
}
