% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_debug.R
\name{vprint}
\alias{vprint}
\title{Print value compactly. Unnest lists, and only show head of data.frames}
\usage{
vprint(x, head_n)
}
\arguments{
\item{x}{value}

\item{head_n}{head n parameter}
}
\description{
Print value compactly. Unnest lists, and only show head of data.frames
}
