% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{ggdebug_one}
\alias{ggdebug_one}
\alias{ggundebug_one}
\title{Debug one ggplot2 function}
\usage{
ggdebug_one(fun, ...)

ggundebug_one(fun, ...)
}
\arguments{
\item{fun}{A function or \code{\link[ggplot2:ggproto]{ggplot2::ggproto()}} method.}

\item{...}{Passed to \code{\link[=debug]{debug()}}, or \code{\link[=debugonce]{debugonce()}}}
}
\description{
Debug one ggplot2 function
}
