% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggmethods}
\alias{ggmethods}
\alias{ggeverything}
\title{Get all methods of a ggproto object}
\usage{
ggmethods(obj, ns = getNamespace("ggplot2"), ignore = "super",
  include_super = TRUE)

ggeverything(ns = getNamespace("ggplot2"), ignore = c("$.ggproto",
  "make_proto_method", "[[.ggproto", "fetch_ggproto", "[[[.ggproto",
  "mean_cl_boot", "median_hilow", "annotation_id", "mean_sdl",
  "mean_cl_normal"))
}
\arguments{
\item{obj}{A ggplot2 ggproto object.}

\item{ns}{Namespace from which objects should be sourced}

\item{ignore}{Method names to ignore}

\item{include_super}{Follow superclass ggproto objects}
}
\value{
A quosures that can be spliced (\code{!!!}) into \code{\link[=ggdebug]{ggdebug()}} or \code{\link[=ggtrace]{ggtrace()}}
}
\description{
Get all methods of a ggproto object
}
