% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{ggtrace_one}
\alias{ggtrace_one}
\alias{gguntrace_one}
\title{Trace one ggplot2 function}
\usage{
ggtrace_one(fun, on_enter = quo(NULL), on_exit = quo(NULL))

gguntrace_one(fun)
}
\arguments{
\item{fun}{A function or \code{\link[ggplot2:ggproto]{ggplot2::ggproto()}} method.}

\item{on_enter}{A \code{\link[rlang:quo]{rlang::quo()}} to evaluate on entering the \code{fun}.}

\item{on_exit}{A \code{\link[rlang:quo]{rlang::quo()}} to evaluate on exiting the \code{fun}.}
}
\description{
Trace one ggplot2 function
}
